Refactor cache clearing logic into `EventCacheHelper`
